Add files via upload
Update Log:
R program accepts terminal input for "username" and "userData.csv".
R program reads and stores the nest location, shape files and srtms in the userData.csv.

Error:
R program was not finding the right path to the nestLocation.csv file. It may be remedied now (check).
'Slope', 'Aspect' and the 'other slope variable' values are not transferring their data to the method prior to changeColNames. Find out why.
R program still performs poorly for the focal() in extractData().